# Conflicts:
#	src/main/java/cot/colabare/profile/controller/LoginController.java
#	src/main/webapp/WEB-INF/views/calendar/calendarform2.jsp
#	src/main/webapp/WEB-INF/views/login/main.jsp
#	target/m2e-wtp/web-resources/META-INF/MANIFEST.MF
#	target/m2e-wtp/web-resources/META-INF/maven/cot.colabare/controller/pom.properties

+++ src/main/java/cot/colabare/approval/mapper/ApproverMapper.java
@@ -21,7 +21,7 @@ public interface ApproverMapper {
 	public ApproverDto approvalNumSelect(Long no);
 	
 	// 내가 받은 결재 목록 
-	public List<ApproverDto> receiveList(Long employee_no);
+	public List<ApproverDto> receiveList(int employee_no);
 	
 	// 반려 
 	public int rejectApver(ApproverDto apverDto);
